loveteppy/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
defusedxml==0.7.1
Django==2.1.5
django-adminlte3==0.1.6
django-allauth==0.44.0
django-bootstrap-modal-forms==2.1.0
django-ckeditor==6.0.0
django-crispy-forms==1.14.0
django-js-asset==1.2.2
idna==3.3
oauthlib==3.2.0
Pillow==8.1.2
pycparser==2.21
PyJWT==2.4.0
PyMySQL==1.0.2
python3-openid==3.2.0
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
urllib3==1.26.11
